ApplicationServer/requirements.txt
Cython
spacy
Pillow
requests
numpy
annoy
pyfunctional
falcon
rethinkdb
gunicorn
futures